Move verifier and downloader files to /persist/vault/agentName/
