+++ pkg/pillar/cmd/downloader/dirs.go
@@ -1,28 +1,24 @@
-// Copyright (c) 2019-2020 Zededa, Inc.
+// Copyright (c) 2020 Zededa, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package verifier
+package downloader
 
 import (
 	"os"
+	"path"
 
 	log "github.com/sirupsen/logrus"
 )
 
-func initializeDirs() {
-	// Remove any files which didn't make it past the verifier.
-	clearInProgressDownloadDirs()
-	// create the object download directories
-	createDownloadDirs()
-}
-
 // Create the object download directories we own
 func createDownloadDirs() {
+
+	workingDirTypes := []string{getPendingDir()}
+
 	// now create the download dirs
-	workingDirTypes := []string{getVerifierDir(), getVerifiedDir()}
 	for _, dirName := range workingDirTypes {
 		if _, err := os.Stat(dirName); err != nil {
-			log.Infof("Create %s", dirName)
+			log.Debugf("Create %s", dirName)
 			if err := os.MkdirAll(dirName, 0700); err != nil {
 				log.Fatal(err)
 			}
@@ -34,7 +30,8 @@ func createDownloadDirs() {
 func clearInProgressDownloadDirs() {
 
 	// Now remove the in-progress dirs
-	workingDirTypes := []string{getVerifierDir()}
+	workingDirTypes := []string{getPendingDir()}
+
 	for _, dirName := range workingDirTypes {
 		if _, err := os.Stat(dirName); err == nil {
 			if err := os.RemoveAll(dirName); err != nil {
@@ -42,4 +39,8 @@ func clearInProgressDownloadDirs() {
 			}
 		}
 	}
+}
+
+func getPendingDir() string {
+	return path.Join(downloaderBasePath, "pending")
 }
